testes para colocar em produção

+++ src/database/knex/migrations/20240908204147_createTags.js
@@ -1,14 +1,17 @@
-exports.up = (knex) =>
-    knex.schema.createTable("tags", (table) => {
-        table.increments("id");
-        table.text("name").notNullable();
+exports.up = async function (knex) {
+    const exists = await knex.schema.hasTable("tags");
+    if (!exists) {
+        return knex.schema.createTable("tags", (table) => {
+            table.increments("id");
+            table.text("name").notNullable();
 
-        table
-            .integer("note_id")
-            .references("id")
-            .inTable("notes")
-            .onDelete("CASCADE");
-        table.integer("user_id").references("id").inTable("users");
-    });
-
-exports.down = (knex) => knex.schema.dropTable("tags");
+            table
+                .integer("note_id")
+                .references("id")
+                .inTable("notes")
+                .onDelete("CASCADE");
+            table.integer("user_id").references("id").inTable("users");
+        });
+    }
+}
+exports.down = (knex) => knex.schema.dropTableIfExists("tags");
